% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{control_summary}
\alias{control_summary}
\title{Summarise dataset for use in control compounds (ctl_compounds) function}
\usage{
control_summary(data, met_vars)
}
\arguments{
\item{data}{tidy dataset from msprep_obj (msprep_obj$data)}
}
\description{
Summarise dataset for use in control compounds (ctl_compounds) function
}
