% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticgen.R
\name{diagnosticgen}
\alias{diagnosticgen}
\title{Generates diagnostic plots for normalization routine}
\usage{
diagnosticgen(testobj, clindat, link1, batch, pheno, directory, ylim1, ylim2,
  ylim3)
}
\arguments{
\item{testobj}{Object output from normdata() function}

\item{clindat}{Name of clinical dataset.}

\item{link1}{Column in clinical dataset that links subject ID to row names
in quantification datasets.}

\item{batch}{Column name for batch identifier}

\item{pheno}{Column name for phenotype identifier}

\item{directory}{Directory to output the diagnostic plots to.}

\item{ylim1}{Limits for median graphs.}

\item{ylim2}{Limits for raw, quant, RUV, and SVA graphs.}

\item{ylim3}{Limits for CRMN graph.}
}
\value{
Placeholder
}
\description{
Generates a PCA plot and three boxplots for each normalization method.
}
\examples{
  load("test.Rdata")
  load("test3.Rdata")
  
  testobj   <- test3
  clindat   <- test$clinical
  link1     <- "SubjectID"
  pheno     <- "Spike"
  batch     <- "Operator"
  directory <- "C:/Users/Hazy/Dropbox/Metabolomics/Programs/MSProcess/"
  ylim2     <- c(10,28)
  ### For median
  ylim1     <- c(-15,15)
  ### for crmn
  ylim3     <- c(18,37)

  diagnosticgen(testobj, clindat, link1, batch, pheno, directory, ylim1,
                ylim2, ylim3)

}
