% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterft.R
\name{filterft}
\alias{filterft}
\title{Filters and imputes dataset}
\usage{
filterft(metaf, filterpercent = 0.5)
}
\arguments{
\item{metaf}{Summarized dataset output as sum_data1 from readdata() function}

\item{filterpercent}{Percent to filter the data}
}
\value{
Placeholder
}
\description{
Filters compounds to those found in specified percentage of subjects and
performs data imputation.
}
\details{
minval Filtered dataset with missing values replaced by 1/2 minimum
observed value for that compound.

bpca Filtered dataset with missing values imputed by a Bayesian PCA
from PCAMethods package.

withzero Filtered dataset with no imputation.

count List of all compounds and the percent present for each
compound.
}
\examples{
  # Load object generated from readdata() function
  load("test.Rdata")
  
  test2 <- filterft(test$sum_data1, 0.80)

}
\references{
Oba, S.et al.(2003) A Bayesian missing value estimation for gene
  expression profile data. Bioinformatics, 19, 2088-2096

  Stacklies, W.et al.(2007) pcaMethods A bioconductor package providing
  PCA methods for incomplete data. Bioinformatics, 23, 1164-1167.
}
