% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphimputations.R
\name{graphimputations}
\alias{graphimputations}
\title{Generate histograms of three datasets returned from filterft function}
\usage{
graphimputations(directory, minval, withzero, bpca, meanval1 = 0,
  meanval2 = 1e+06, xmax1 = 5e+05, ymax1 = 25000, xmax2 = 20,
  ymax2 = 1000, xmax3 = 20, ymax3 = 1500, nbreaks = 500)
}
\arguments{
\item{directory}{Location to save pdf of histograms.}

\item{minval}{Name of minval dataset}

\item{withzero}{Name of withzero dataset}

\item{bpca}{Name of BPCA dataset}

\item{meanval1}{Minimum value to keep.}

\item{meanval2}{Maximum value to keep.}

\item{xmax1}{X limit for withzero graph}

\item{ymax1}{Y limit for withzero graph}

\item{xmax2}{X limit for minval graph}

\item{ymax2}{Y limit for minval graph}

\item{xmax3}{X limit for bpca graph}

\item{ymax3}{Y limit for bpca graph}

\item{nbreaks}{Number of breaks in histogram}
}
\value{
Placeholder
}
\description{
Generates histograms of withzero, minval, and bpca datasets.
}
\details{
The meanval1 and meanval2 allow removal of any individual extreme
values or ability to generate a focused histogram of abundances of interest.
nbreaks will need increased as the number of subjects*compounds increases.
}
\examples{
  load("test2.Rdata")
  
  directory <- "C:/Users/Hazy/Dropbox/Metabolomics/Programs/MSProcess/"
  minval    <- test2$minval
  withzero  <- test2$withzero
  bpca      <- test2$bpca
  
  graphimputations(directory, minval, withzero, bpca,fname, meanval1 = 0,
                   meanval2 = 200000, xmax1 = 400000, ymax1 = 800, 
                   xmax2 = 20, ymax2 = 600, xmax3 = 20, ymax3 = 175, 
                   nbreaks = 200)


}
