% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{ms_filter}
\alias{ms_filter}
\title{Function for filtering prepared dataset.}
\usage{
ms_filter(msprep_obj, filter_percent = 0.5)
}
\arguments{
\item{msprep_obj}{Prepared MSPrep object.}

\item{filter_percent}{Decimal value representing to proportion to filter the data.}
}
\value{
An `msprep` object with `stage(rtn) == "filtered"` containing
filtered quantification data.
}
\description{
Filters compounds to those found in specified proportion of subjects.
}
\examples{

library(magrittr)

# Load example dataset, tidy it, prepare it, then filter it
data(msquant)

prepped_data <- msquant \%>\% ms_tidy \%>\%
  ms_prepare(replicate = "replicate",
             batch = "batch",
             groupingvars = "spike")
filtered_data <- ms_filter(prepped_data, 0.80)

}
\references{
Oba, S.et al.(2003) A Bayesian missing value estimation for gene
  expression profile data. Bioinformatics, 19, 2088-2096

  Stacklies, W.et al.(2007) pcaMethods A bioconductor package providing
  PCA methods for incomplete data. Bioinformatics, 23, 1164-1167.
}
