% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{ms_impute}
\alias{ms_impute}
\title{Function for imputing on filtered data.}
\usage{
ms_impute(msprep_obj, imputeMethod = c("halfmin", "bpca", "knn"),
  k_knn = 5, n_pcs = 3, compoundsAsNeighbors = FALSE)
}
\arguments{
\item{msprep_obj}{Filtered MSPrep object.}

\item{imputeMethod}{Name of imputation method to use. 
Options are:
- halfmin (half the minimum value)
- bpca (Bayesian PCA)
- knn (k-nearest neighbors)}

\item{k_knn}{Number of clusters for 'knn' method.}

\item{n_pcs}{Number of  principle components used for re-estimation for 
'bpca' method.}

\item{compoundsAsNeighbors}{If TRUE, will use compounds as neighbors for KNN imputation
rather than samples. Note, using compounds as neighbors is significantly slower than using
samples as neighbors.}
}
\value{
An msprep object with missing data imputed.
}
\description{
Performs data imputation for a given mz_rt value (not separately for batches
or grouping variables).
}
\details{
minval Filtered dataset with missing values replaced by 1/2 minimum
observed value for that compound.

bpca Filtered dataset with missing values imputed by a Bayesian PCA
from PCAMethods package.
}
\examples{
library(magrittr)

# Load object generated from readdata() function
data(msquant)

filtered_data <- msquant \%>\% ms_tidy \%>\% 
  ms_prepare(replicate = "replicate", 
             batch = "batch", 
             groupingvars = "spike") \%>\% 
  ms_filter(0.80)
imputed_data  <- ms_impute(filtered_data, imputeMethod = "halfmin")

}
\references{
Oba, S.et al.(2003) A Bayesian missing value estimation for gene
  expression profile data. Bioinformatics, 19, 2088-2096

  Stacklies, W.et al.(2007) pcaMethods A bioconductor package providing
  PCA methods for incomplete data. Bioinformatics, 23, 1164-1167.
}
