% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{ms_prepare}
\alias{ms_prepare}
\alias{print.msprepped}
\alias{replace_missing}
\alias{select_summary_measure}
\alias{standardize_dataset}
\title{Prepare a mass spec quantification data frame for filtering, imputation,
and normalization. Also provides summaries of data structure (replicates,
subjects, spike, etc.)}
\usage{
ms_prepare(data, subject_id = "subject_id", replicate = NULL,
  abundance = "abundance", spike = "spike", mz = "mz", rt = "rt",
  cvmax = 0.5, missing_val = 1, min_proportion_present = 1/3)

\method{print}{msprepped}(x)

replace_missing(abundance, missing_val)

select_summary_measure(n_present, cv_abundance, n_replicates,
  min_proportion_present, cv_max)

standardize_dataset(data, subject_id, replicate, abundance, spike, mz, rt)
}
\arguments{
\item{data}{A tidy dataframe of quantification data.}

\item{mz}{Mass-to-charge ratio variable name.}

\item{rt}{Retention time variable name.}

\item{cvmax}{Acceptable level of coefficient of variation between replicates.}

\item{missing}{Value of missing data in the quantification data file.}

\item{linktxt}{Column name for Run ID field in Subject Link dataset}
}
\value{
An \code{msprepped} object containing summarised quantification data,
a dataset of compounds summarised by medians, and other related summaries.
}
\description{
Function reads in raw data files and summarizes technical replicates as the
mean of observations for compounds found in 2 or 3 replicates and with
coefficient of variation below specified level, or median for those found in
3 replicates but excess CV.
}
\examples{

# Read in data file
data(msquant)

# Convert dataset to tidy format
tidy_data    <- ms_tidy(msquant, mz = "mz", rt = "rt")
prepped_data <- ms_prepare(tidy_data, replicate = "replicate")

# Or, using tidyverse/magrittr pipes 
prepped_data <- msquant \%>\% ms_tidy \%>\% ms_prepare

str(prepped_data)
str(prepped_data$summary_data)

}
