% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return.R
\name{ms_return}
\alias{ms_return}
\title{Function to return data prepared by MSPrep pipeline to compound by sample format}
\usage{
ms_return(msprep_obj)
}
\arguments{
\item{msprep_obj}{An object of class `msprep`}
}
\description{
Function to return data prepared by MSPrep pipeline to compound by sample format
}
