% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{ms_summarize}
\alias{ms_summarize}
\title{Function for summarizing tidied dataset and preparing for filtering, imputation, 
and normalization.}
\usage{
ms_summarize(data, abundance = "abundance", met_id = NULL, mz = NULL,
  rt = NULL, subject_id = "subject_id", replicate = NULL,
  batch = NULL, groupingvars = NULL, cvmax = 0.5, missing_val = 1,
  min_proportion_present = 1/3)
}
\arguments{
\item{data}{Tidied dataset.}

\item{abundance}{Name of the abundance column.}

\item{met_id}{Name of compound name variable.}

\item{mz}{Name of mass-to-charge ratio variable.}

\item{rt}{Name of retention time variable.}

\item{subject_id}{Name of the subject ID column.}

\item{replicate}{Name of the replicate column. Set to NULL if no
replicates.}

\item{batch}{Name of the column representing batches. Set to NULL if no batches present.}

\item{groupingvars}{Variable name or vector of names of the
phenotypes or comparison groups. Set to NULL if none are present.}

\item{cvmax}{Decimal value from 0 to 1 representing the acceptable level of coefficient 
of variation between replicates.}

\item{missing_val}{Value of missing data in the quantification data file.}

\item{min_proportion_present}{Decimal value from 0 to 1 representing the minimum proportion present 
to summarize with median or mean. Below this the compound will be set to 0.}
}
\value{
An `msprep` object with `stage(rtn) == "prepared"` containing
summarised quantification data, a dataset of compounds summarised by medians,
and other related summaries.
}
\description{
Prepares a mass spec quantification data frame for filtering, imputation,
and normalization. Also provides summaries of data structure (replicates,
subjects, groupingvars, etc.)
}
\details{
Function reads in raw data files and summarizes technical replicates as the
mean of observations for compounds found in 2 or 3 replicates and with
coefficient of variation below specified level, or median for those found in
3 replicates but excess CV.
}
\examples{

# Read in data file
data(msquant)

# Convert dataset to tidy format
tidy_data    <- ms_tidy(msquant, mz = "mz", rt = "rt")
prepped_data <- ms_summarize(tidy_data, 
                           replicate = "replicate", 
                           batch = "batch",
                           groupingvars = "spike")

# Or, using tidyverse/magrittr pipes 
library(magrittr)
prepped_data <- msquant \%>\% ms_tidy \%>\%
  ms_summarize(replicate = "replicate",
             batch = "batch",
             groupingvars = "spike")

str(prepped_data)
str(prepped_data$summary_data)

}
