% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{ms_tidy}
\alias{ms_tidy}
\title{Function for converting wide mass spectrometry quantification data into a tidy data
frame.}
\usage{
ms_tidy(quantification_data, mz = "mz", rt = "rt",
  col_extra_txt = "Neutral_Operator_Dif_Pos_", col_names = c("spike",
  "batch", "replicate", "subject_id"), separator = "_")
}
\arguments{
\item{quantification_data}{Data frame containing the quantification data.}

\item{mz}{Name of the column containing mass-to-charge ratios.}

\item{rt}{Name of the column containing retention time.}

\item{col_extra_txt}{Text to remove from column names when converting column names to
variables.}

\item{col_names}{Vector of the ordered ID names to extract from the variable
names.}

\item{separator}{Character or text separating spike, subject, and replicate
ids in column names.}
}
\value{
A tidy data frame of quant data, with columns mz, rt,
replicate, and abundance.
}
\description{
Function reads in wide dataset of mass spectrometry quantification data and converts
it to a tidy dataset.  This function assumes that the dataset is in a wide
format, with a column representing the retention time (rt), another
representing the mass-to-charge ratio (mz), and the remaining columns
containing MS quantification data.
}
\details{
It also assumes that the column names of quantification data start with some
consistent, informational but unnecessary text (col_extra_txt), and contain
the spike, subject ID, and replicate ID in a consistent position, all
separated by a consistent separator.

See \code{data(quantification)} for an example.  If your data doesn't fit
this format, view the function code for hints on tidying your data.  The core
of this function consists of \code{tidyr::gather()}, \code{dplyr::mutate()},
and \code{tidyr::separate()}.
}
\examples{

  # load raw abundance/quantification dataset
  data(msquant)

  # print original dataset (using tibble's nice printing)
  tibble::as_data_frame(msquant) 

  # tidy the dataset and print it
  ms_tidy(msquant, mz = "mz", rt = "rt")

}
