% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msquant.R
\docType{data}
\name{msquant_subject1}
\alias{msquant_subject1}
\title{Example quantification dataset}
\format{\code{
'data.frame':   2654 obs. of  29 variables:
 $ mz                              : num  577 539 724 525 483 ...
 $ rt                              : Factor w/ 2510 levels "0.481","0.58933336",..: 3 2 6 15 10 8 22 18 17 13 ...
 $ Neutral_Operator_Dif_Pos_1x_O1_A_01: num  1 1 36270 105305 1 ...
 $ Neutral_Operator_Dif_Pos_1x_O1_B_01: num  1 1 35271 92489 1 ...
 $ Neutral_Operator_Dif_Pos_1x_O1_C_01: num  1 1 39355 105189 1 ...
 $ Neutral_Operator_Dif_Pos_1x_O2_A_01: num  24182 20686 57021 141999 67760 ...
 $ Neutral_Operator_Dif_Pos_1x_O2_B_01: num  25857 24500 60641 151657 73151 ...
 $ Neutral_Operator_Dif_Pos_1x_O2_C_01: num  24616 22820 61023 144739 71789 ...
}}
\usage{
msquant_subject1
}
\description{
Mass spec data example with 1 subject
}
\details{
mz = Mass/Charge ratio
rt = Retention time

The indicators in the column names are spike-in (e.g. 1x), batch (e.g. 01),
technical replicate (e.g. A), and subject ID (e.g. 01).
}
\examples{
  data(msquant_subject1)
  str(msquant_subject1)
}
\keyword{datasets}
