% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_combat}
\alias{normalize_combat}
\title{Apply only batch correction via ComBat}
\usage{
normalize_combat(data, groupingvars, batch)
}
\arguments{
\item{data}{The 'data' element from MSPrep object (msprep_obj).}

\item{groupingvars}{The groupingvars element from MSPrep object, aka
phenotypes.}

\item{batch}{The 'batch' element from MSPrep object. Name for batch variable}
}
\description{
Apply only batch correction via ComBat
}
