% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R
\name{readdata}
\alias{readdata}
\title{Function for importing raw data and summarizing replicates}
\usage{
readdata(directory, clinicalfile, quantificationfile, linkfile, cvmax = 0.5,
  missing = 1, linktxt)
}
\arguments{
\item{clinicalfile}{Name of the clinical data file.}

\item{quantificationfile}{Name of the data file containing the
quantification data.}

\item{linkfile}{Name of Subject link file.}

\item{cvmax}{Acceptable level of coefficient of variation between replicates.}

\item{missing}{Value of missing data in the quantification data file.}

\item{linktxt}{Column name for Run ID field in Subject Link dataset}

\item{Directory}{where the required three data files are located.}
}
\value{
sum_data1 Matrix of summarized replicates, one obs per subject per
compound

clinical Clinical dataset

medians List of compounds that had excess CV and utilized the median
}
\description{
Function reads in raw data files and summarizes technical replicates as the
mean of observations for compounds found in 2 or 3 replicates and with
coefficient of variation below specied level, or median for those found in 3
replicates but excess CV.
}
\note{
Must have unique compounds in raw data file.  Raw quantification data
should have two columns titled mz and rt that are combined to make the
column header.
}
\examples{
  ### Specify primary directory
  directory <- c("/MSProcess/data/")
  
  ### Specify location of data files
  clinicalfile       <- c("Clinical.csv")
  quantificationfile <- c("Quantification.csv")
  linkfile <- c("SubjectLinks.csv")
  
  ### Set variables for program
  cvmax   <- 0.5
  missing <- 1
  linktxt <- "LCMS_Run_ID"
  
  test <- readdata(directory, clinicalfile, quantificationfile, linkfile,
                   cvmax = 0.50, missing = 1, linktxt)
  save(test, file = paste(directory, "test.Rdata", sep = ""))

}
