% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test2_data.R
\docType{data}
\name{test2}
\alias{test2}
\title{Object exported from the filterft() function.}
\format{The format is:
  List of 4
  $ minval  :'data.frame':	9 obs. of  891 variables:
   ..$ 577.0322_0.5910416  : num [1:9] 9445 24885 23820 20730 19302 ...
   ..$ 539.0207_0.58933336 : num [1:9] 8035 22669 20792 19486 16070 ...
   ..$ 723.9585_0.599889   : num [1:9] 36965 59562 58157 46326 46133 ...
   ..$ 525.3529_0.6085925  : num [1:9] 100994 146132 148889 118769 114334 ...
   ..$ 404.2655_0.6037692  : num [1:9] 30093 51270 46124 40558 36139 ...
   .. [list output truncated]
  $ count   :'data.frame':	2654 obs. of  1 variable:
   ..$ V1: num [1:2654] 1 1 0 0 4 0 0 0 0 0 ...}
\usage{
test2
}
\description{
Object exported from the filterft() function.  Contains filtered and
imputation datasets
}
\examples{
  data(test2)
  str(test2)
}
\keyword{datasets}
