% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test3_data.R
\docType{data}
\name{test3}
\alias{test3}
\title{Object exported from the normdata() function.}
\format{The format is:
  List of 12
   $ log_data        :'data.frame':	9 obs. of  891 variables:
    ..$ 577.0322_0.5910416  : num [1:9] 14.4 14.6 14.5 14.3 14.2 ...
    ..$ 539.0207_0.58933336 : num [1:9] 14.2 14.5 14.3 14.3 14 ...
    ..$ 723.9585_0.599889   : num [1:9] 15.2 15.9 15.8 15.5 15.5 ...
    ..$ 525.3529_0.6085925  : num [1:9] 16.6 17.2 17.2 16.9 16.8 ...
    ..$ 404.2655_0.6037692  : num [1:9] 14.9 15.6 15.5 15.3 15.1 ...
    ..$ 1164.7386_0.6163333 : num [1:9] 14.7 15.4 15.4 14.9 15 ...
    ..$ 333.722_0.6112592   : num [1:9] 16.2 17 16.9 16.6 16.6 ...
    .. [list output truncated]
   $ log_data_combat : num [1:9, 1:891] 14.3 14.5 14.5 14.3 14.2 ...
    ..- attr(*, "dimnames")=List of 2
    .. ..$ : chr [1:9] "1x_O1" "1x_O2" "1x_O3" "2x_O1" ...
    .. ..$ : chr [1:891] "577.0322_0.5910416" "539.0207_0.58933336" ...
   $ log_quant       :'data.frame':	9 obs. of  891 variables:
    ..$ 577.0322_0.5910416  : num [1:9] 14.3 14.7 14.5 14.4 14 ...
    ..$ 539.0207_0.58933336 : num [1:9] 14.2 14.6 14.3 14.3 13.6 ...
    ..$ 723.9585_0.599889   : num [1:9] 15.2 15.9 15.8 15.5 15.4 ...
    ..$ 525.3529_0.6085925  : num [1:9] 16.6 17.2 17.1 16.9 16.7 ...
    ..$ 404.2655_0.6037692  : num [1:9] 14.9 15.7 15.5 15.3 15 ...
    ..$ 1164.7386_0.6163333 : num [1:9] 14.8 15.5 15.4 14.9 14.9 ...
    .. [list output truncated]
   $ log_quant_combat: num [1:9, 1:891] 14.2 14.6 14.5 14.4 14.1 ...
    ..- attr(*, "dimnames")=List of 2
    .. ..$ : chr [1:9] "1x_O1" "1x_O2" "1x_O3" "2x_O1" ...
    .. ..$ : chr [1:891] "577.0322_0.5910416" "539.0207_0.58933336" ... 
   $ med_adj         :'data.frame':	9 obs. of  881 variables:
    ..$ 577.0322_0.5910416  : num [1:9] -2.17 -1.94 -2.05 -2.23 -2.39 ...
    ..$ 539.0207_0.58933336 : num [1:9] -2.35 -2.07 -2.24 -2.32 -2.65 ...
    ..$ 723.9585_0.599889   : num [1:9] -1.404 -0.677 -0.759 -1.066 -1.131 ...
    ..$ 525.3529_0.6085925  : num [1:9] 0.0463 0.6174 0.5968 0.2925 0.178 ...
    ..$ 404.2655_0.6037692  : num [1:9] -1.701 -0.894 -1.094 -1.258 -1.484 ...
    .. [list output truncated]
   $ med_combat      : num [1:9, 1:881] -2.22 -2.02 -2.07 -2.24 -2.38 ...
    ..- attr(*, "dimnames")=List of 2
    .. ..$ : chr [1:9] "1x_O1" "1x_O2" "1x_O3" "2x_O1" ...
    .. ..$ : chr [1:881] "577.0322_0.5910416" "539.0207_0.58933336" ...
   $ sva_factors     : num [1:9, 1] -0.1567 0.0432 -0.1536 -0.2078 0.5435 ...
    ..- attr(*, "dimnames")=List of 2
    .. ..$ : NULL
    .. ..$ : chr "f1"
   $ sva_adj         : num [1:9, 1:891] 14.4 14.4 14.4 14.5 14.3 ...
    ..- attr(*, "dimnames")=List of 2
    .. ..$ : chr [1:9] "1x_O1" "1x_O2" "1x_O3" "2x_O1" ...
    .. ..$ : chr [1:891] "577.0322_0.5910416" "539.0207_0.58933336" ...
   $ ruv_factors     : num [1:9, 1] 0.1616 -0.7965 -0.1041 0.0635 0.1234 ...
    ..- attr(*, "dimnames")=List of 2
    .. ..$ : NULL
    .. ..$ : chr "f1"
   $ ruv_adj         : num [1:9, 1:891] 14.4 14.5 14.4 14.4 14.4 ...
    ..- attr(*, "dimnames")=List of 2
    .. ..$ : chr [1:9] "1x_O1" "1x_O2" "1x_O3" "2x_O1" ...
    .. ..$ : chr [1:891] "577.0322_0.5910416" "539.0207_0.58933336" ...
   $ crmn_adj        :'data.frame':	9 obs. of  881 variables:
    ..$ 577.0322_0.5910416  : num [1:9] 20.8 21 21 20.7 20.5 ...
    ..$ 539.0207_0.58933336 : num [1:9] 20.6 20.8 20.7 20.5 20.2 ...
    ..$ 723.9585_0.599889   : num [1:9] 22.2 22.6 22.8 22.2 22.4 ...
    ..$ 525.3529_0.6085925  : num [1:9] 24.2 24.6 24.7 24.2 24.3 ...
    ..$ 404.2655_0.6037692  : num [1:9] 21.8 22.3 22.3 22 21.9 ...
    ..$ 1164.7386_0.6163333 : num [1:9] 21.5 22 22.2 21.4 21.7 ...
    .. [list output truncated]
   $ controls        : chr [1:10, 1:2] "610" "719" "734" "723" ...
    ..- attr(*, "dimnames")=List of 2
    .. ..$ : NULL
    .. ..$ : chr [1:2] "ctl" ""}
\usage{
test3
}
\description{
Object exported from the normdata() function.
}
\examples{
  data(test3)
  str(test3)
}
\keyword{datasets}
