% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COPD_131.R
\docType{data}
\name{COPD_131}
\alias{COPD_131}
\title{Example mass spectrometry dataset}
\format{An object of class \code{data.frame} with 662 rows and 396 columns.}
\usage{
COPD_131
}
\description{
Size: 662 obs. of 396 variables
}
\details{
Data contains samples for 131 subjects with 3 technical replicates per
subject. Additionally, the first three variables are "Mass" which
is the mass-to-charge ratio for each metabolite, "Retention.Time", and
"Compound.Name".
}
\examples{
data(COPD_131)
}
\keyword{datasets}
