% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{ms_filter}
\alias{ms_filter}
\title{Function for filtering prepared dataset.}
\usage{
ms_filter(msprep_obj, filter_percent = 0.8)
}
\arguments{
\item{msprep_obj}{Summarized MSPrep object.}

\item{filter_percent}{Decimal value indicating filtration threshold. 
Metabolites which are present in fewer samples than the specified proportion 
will be removed.}
}
\value{
An `msprep` object with containing filtered quantification data.
}
\description{
Filters compounds to those found in specified proportion of subjects.
}
\examples{

# Load example dataset, tidy it, and summarize it
data(msquant)

tidied_data <- ms_tidy(msquant, mz = 'mz', rt = 'rt',
                       col_extra_txt = 'Neutral_Operator_Dif_Pos_',
                       separator = '_', 
                       col_names = c('spike', 'batch', 'replicate', 
                       'subject_id'))

summarized_data <- ms_summarize(tidied_data, 
                                mz = 'mz', 
                                rt = 'rt', 
                                replicate = 'replicate', 
                                batch = 'batch', 
                                groupingvars = 'spike', 
                                subject_id = 'subject_id', 
                                cvmax = 0.50, 
                                min_proportion_present = 1/3, 
                                missing_val = 1)

# Filter the dataset using a 80\% filter rate
filtered_data <- ms_filter(summarized_data, 
                          filter_percent =  0.80)
                          
# Print results
print(filtered_data)

}
\references{
Oba, S.et al.(2003) A Bayesian missing value estimation for gene
  expression profile data. Bioinformatics, 19, 2088-2096

  Stacklies, W.et al.(2007) pcaMethods A bioconductor package providing
  PCA methods for incomplete data. Bioinformatics, 23, 1164-1167.
}
