% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{ms_impute}
\alias{ms_impute}
\title{Imputes dataset}
\usage{
ms_impute(msprepped, impute_function = ~minval(.x))
}
\arguments{
\item{msprepped}{Placeholder}

\item{impute_function}{Placeholder}
}
\value{
Placeholder
}
\description{
and
performs data imputation.
}
\details{
minval Filtered dataset with missing values replaced by 1/2 minimum
observed value for that compound.

bpca Filtered dataset with missing values imputed by a Bayesian PCA
from PCAMethods package.

withzero Filtered dataset with no imputation.

count List of all compounds and the percent present for each
compound.
}
\examples{
  # Load object generated from readdata() function
  load("./data/test.rda")

  test2 <- ms_filter(test$sum_data1, 0.80)

}
\references{
Oba, S.et al.(2003) A Bayesian missing value estimation for gene
  expression profile data. Bioinformatics, 19, 2088-2096

  Stacklies, W.et al.(2007) pcaMethods A bioconductor package providing
  PCA methods for incomplete data. Bioinformatics, 23, 1164-1167.
}
