% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{ms_impute}
\alias{ms_impute}
\title{Imputes dataset}
\usage{
ms_impute(msprep_obj, method = c("halfmin", "bpca", "knn"))
}
\arguments{
\item{msprep_obj}{Placeholder}

\item{method}{Placeholder}
}
\value{
Placeholder
}
\description{
Performs data imputation.
}
\details{
minval Filtered dataset with missing values replaced by 1/2 minimum
observed value for that compound.

bpca Filtered dataset with missing values imputed by a Bayesian PCA
from PCAMethods package.

withzero Filtered dataset with no imputation.

count List of all compounds and the percent present for each
compound.
}
\examples{
library(magrittr)

# Load object generated from readdata() function
data(msquant)

filtered_data <- msquant \%>\% ms_tidy \%>\% ms_prepare \%>\% ms_filter(0.80)
imputed_data <- ms_impute(filtered_data, "halfmin")

}
\references{
Oba, S.et al.(2003) A Bayesian missing value estimation for gene
  expression profile data. Bioinformatics, 19, 2088-2096

  Stacklies, W.et al.(2007) pcaMethods A bioconductor package providing
  PCA methods for incomplete data. Bioinformatics, 23, 1164-1167.
}
