% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{ms_impute}
\alias{ms_impute}
\title{Imputes dataset}
\usage{
ms_impute(msprep_obj, method = c("halfmin", "bpca", "knn"), k = 5)
}
\arguments{
\item{msprep_obj}{An MSPrep object that has been filtered and normalized}

\item{method}{Imputation method.  One of half-min (half the minimum value),
bpca (Bayesian PCA), knn (k-nearest neighbors).}

\item{k}{Number of clusters for 'knn' method.}
}
\value{
A msprep object with missing data imputed.
}
\description{
Performs data imputation.
}
\details{
minval Filtered dataset with missing values replaced by 1/2 minimum
observed value for that compound.

bpca Filtered dataset with missing values imputed by a Bayesian PCA
from PCAMethods package.
}
\examples{
library(magrittr)

# Load object generated from readdata() function
data(msquant)

filtered_data <- msquant \%>\% ms_tidy \%>\% ms_prepare \%>\% ms_filter(0.80)
imputed_data  <- ms_impute(filtered_data, "bpca")

}
\references{
Oba, S.et al.(2003) A Bayesian missing value estimation for gene
  expression profile data. Bioinformatics, 19, 2088-2096

  Stacklies, W.et al.(2007) pcaMethods A bioconductor package providing
  PCA methods for incomplete data. Bioinformatics, 23, 1164-1167.
}
