% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{ms_impute}
\alias{ms_impute}
\title{Function for imputing missing values in data.}
\usage{
ms_impute(
  msprep_obj,
  imputeMethod = c("halfmin", "bpca", "knn"),
  k_knn = 5,
  n_pcs = 3,
  compoundsAsNeighbors = FALSE
)
}
\arguments{
\item{msprep_obj}{Filtered MSPrep object.}

\item{imputeMethod}{String specifying imputation method choice. 
Options are "halfmin" (half the minimum value), "bpca" (Bayesian PCA), 
and "knn" (k-nearest neighbors)}

\item{k_knn}{Number of clusters for 'knn' method.}

\item{n_pcs}{Number of  principle components used for re-estimation for 
'bpca' method.}

\item{compoundsAsNeighbors}{For KNN imputation. If TRUE, compounds will be 
used as neighbors rather than samples. Note that using compounds as 
neighbors is significantly slower than using samples.}
}
\value{
An `msprep` object with missing data imputed.
}
\description{
Replaces missing values with non-zero estimates calculated using a
selected method.
}
\examples{
# Load, tidy, summarize, and filter example dataset
data(msquant)

tidied_data <- ms_tidy(msquant, mz = "mz", rt = "rt",
                       col_extra_txt = "Neutral_Operator_Dif_Pos_",
                       separator = "_", 
                       col_names = c("spike", "batch", "replicate", 
                                     "subject_id"))

summarized_data <- ms_summarize(tidied_data, 
                                mz = "mz", 
                                rt = "rt", 
                                replicate = "replicate", 
                                batch = "batch", 
                                groupingvars = "spike", 
                                subject_id = "subject_id", 
                                cvmax = 0.50, 
                                min_proportion_present = 1/3, 
                                missing_val = 1)

filtered_data <- ms_filter(summarized_data, 
                           filter_percent =  0.80)
                           
# Impute dataset using 3 possible options
imputed_data_hm <- ms_impute(filtered_data, 
                             imputeMethod = "halfmin")
imputed_data_knn <- ms_impute(filtered_data, 
                              imputeMethod = "knn",
                              k_knn = 5)
imputed_data_bpca <- ms_impute(filtered_data,
                               imputeMethod = "bpca",
                               n_pcs = 3)

}
\references{
Oba, S.et al.(2003) A Bayesian missing value estimation for gene
  expression profile data. Bioinformatics, 19, 2088-2096

  Stacklies, W.et al.(2007) pcaMethods A bioconductor package providing
  PCA methods for incomplete data. Bioinformatics, 23, 1164-1167.
}
