% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{ms_normalize}
\alias{ms_normalize}
\title{Function for performing normalization and batch corrections on imputed data.}
\usage{
ms_normalize(msprep_obj, normalizeMethod = c("median", "ComBat",
  "quantile", "quantile + ComBat", "median + ComBat", "CRMN", "RUV",
  "SVA"), n_control = 10, controls = NULL, n_comp = 2, k_ruv = 3,
  transform = c("log10", "log2", "none"))
}
\arguments{
\item{msprep_obj}{Imputed MSPrep object.}

\item{normalizeMethod}{Name of normalization method.
ComBat" (only ComBat batch correction), "quantile" (only quantile 
normalization), "quantile + ComBat" (quantile with ComBat batch correction),
"median" (only median normalization), "median + ComBat" (median with ComBat
batch correction), "CRMN" (cross-contribution compensating multiple 
standard normalization), "RUV" (remove unwanted variation), "SVA" (surrogate 
variable analysis)}

\item{n_control}{Number of controls to estimate/utilize.}

\item{controls}{Vector of control identifiers.  Leave blank for data driven
controls. Vector of column numbers from metafin dataset of that control.}

\item{n_comp}{Number of factors to use in CRMN algorithm.}

\item{k_ruv}{Number of factors to use in RUV algorithm.}

\item{transform}{Select transformation to apply to data prior to normalization.
- log10
- log2
- none}
}
\value{
controls List of compounds that were used as controls.

crmn_adj Log2 CRMN normalized data.

log_data Log2 dataset (no normalization)

log_data_combat Log2 with ComBat batch effect corrected dataset.

log_quant Log2 quantile normalized dataset.

log_quant_combat Log2 quantile with ComBat correction normalized dataset.

med_adj Median normalized dataset.

med_combat Median with ComBat normalized dataset.

ruv_adj Log2 RUV normalized dataset .

ruv_factors Matrix of factors generated in RUV for use as covariates
in later analysis.

sva_adj Log2 SVA normalized dataset.

sva_factors Matrix of factors generated in SVA for use as covariates
in later analysis.
}
\description{
Perform normalization and batch corrections on specified imputation dataset.
Routines included are quantile, RUV (remove unwanted variation), SVA 
(surrogate variable analysis), median, CRMN (cross-contribution 
compensating multiple standard normalization), ComBat to remove batch 
effects in raw, quantile, and median normalized data. Generates data 
driven controls if none exist.
}
\examples{
# Load, tidy, summarize, filter, and impute example dataset
data(msquant)

tidied_data <- ms_tidy(msquant, mz = "mz", rt = "rt",
                       col_extra_txt = "Neutral_Operator_Dif_Pos_",
                       separator = "_", 
                       col_names = c("spike", "batch", "replicate", "subject_id"))

summarized_data <- ms_summarize(tidied_data, 
                                mz = "mz", 
                                rt = "rt", 
                                replicate = "replicate", 
                                batch = "batch", 
                                groupingvars = "spike", 
                                subject_id = "subject_id", 
                                cvmax = 0.50, 
                                min_proportion_present = 1/3, 
                                missing_val = 1)

filtered_data <- ms_filter(summarized_data, 
                           filter_percent =  0.80)

imputed_data <- ms_impute(filtered_data, 
                          imputeMethod = "halfmin")

# Normalize dataset using six possible options
normalized_data_median <- ms_normalize(imputed_data,
                                       normalizeMethod = "median")
normalized_data_combat <- ms_normalize(imputed_data,
                                       normalizeMethod = "ComBat")
normalized_data_quantile <- ms_normalize(imputed_data,
                                         normalizeMethod = "quantile")
normalized_data_crmn <- ms_normalize(imputed_data,
                                     normalizeMethod = "CRMN")
normalized_data_ruv <- ms_normalize(imputed_data,
                                    normalizeMethod = "RUV")
normalized_data_sva <- ms_normalize(imputed_data,
                                    normalizeMethod = "SVA")


}
\references{
Bolstad, B.M.et al.(2003) A comparison of normalization methods for high
density oligonucleotide array data based on variance and bias.
Bioinformatics, 19, 185-193

DeLivera, A.M.et al.(2012) Normalizing and Integrating Metabolomic Data.
Anal. Chem, 84, 10768-10776.

Gagnon-Bartsh, J.A.et al.(2012) Using control genes to correct for unwanted
variation in microarray data. Biostatistics, 13, 539-552.

Johnson, W.E.et al.(2007) Adjusting batch effects in microarray expression
data using Empirical Bayes methods. Biostatistics, 8, 118-127.

Leek, J.T.et al.(2007) Capturing Heterogeneity in Gene Expression Studies by
Surrogate Variable Analysis. PLoS Genetics, 3(9), e161

Wang, W.et al.(2003) Quantification of Proteins and Metabolites by Mass
Spectrometry without Isotopic Labeling or Spiked Standards. Anal. Chem., 75,
4818-4826.
}
