% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{ms_normalize}
\alias{ms_normalize}
\title{Perform normalization and batch corrections}
\usage{
ms_normalize(msprep_obj, method = c("ComBat", "quantile + ComBat",
  "median + ComBat", "CRMN", "RUV", "SVA"), n_control = 10,
  controls = NULL, n_comp = 2)
}
\arguments{
\item{msprep_obj}{An MSPrep object that has been filtered, normalized, and
imputed.}

\item{method}{One of 6 methods.  3 are normalization and 3 are batch
correction + normalization? TODO: updated this.}

\item{n_control}{Number of controls to estimate/utilize.}

\item{controls}{Vector of control identifiers.  Leave blank for data driven
controls. Vector of column numbers from metafin dataset of that control.}

\item{n_comp}{Number of factors to use in CRMN algorithm.}
}
\value{
controls List of compounds that were used as controls

crmn_adj CRMN adjusted dataset (log2)

log_data Log2 data with no adjustment

log_data_combat Log2 data with Combat batch adjustment

log_quant Log2 Quantile adjusted

log_quant_combat Log2 Quantile with Combat data

med_adj Median adjusted dataset

med_combat Median adjusted with Combat

ruv_adj RUV adjusted dataset (log2)

ruv_factors Matrix of factors generated in RUV for use as covariates
in later analysis.

sva_adj SVA adjusted dataset (log2)

sva_factors Matrix of factors generated in SVA for use as covariates
in later analysis.
}
\description{
Perform normalization and batch corrections on specified imputation dataset.
Routines included are quantile, RUV, SVA, median, and CRMN.  Combat to
remove batch effects in raw, quantile, and median normalized data.
Generates data driven controls if none exist.
}
\examples{
library(magrittr)

# Load object generated from readdata() function
data(msquant)

imputed_data <-
  msquant \%>\%
  ms_tidy \%>\%
  ms_prepare(replicate = "replicate", batch = "batch", groupingvars = "spike") \%>\%
  ms_filter(filter_percent = 0.80) \%>\%
  ms_impute(method = "halfmin")
normalized_data_qC   <- ms_normalize(imputed_data, method = "quantile + ComBat")
normalized_data_C    <- ms_normalize(imputed_data, method = "ComBat")
normalized_data_sva  <- ms_normalize(imputed_data, method = "SVA")
normalized_data_ruv  <- ms_normalize(imputed_data, method = "RUV")
normalized_data_crmn <- ms_normalize(imputed_data, method = "CRMN")
normalized_data_medC <- ms_normalize(imputed_data, method = "median + ComBat")


}
\references{
Bolstad, B.M.et al.(2003) A comparison of normalization methods for high
density oligonucleotide array data based on variance and bias.
Bioinformatics, 19, 185-193

DeLivera, A.M.et al.(2012) Normalizing and Integrating Metabolomic Data.
Anal. Chem, 84, 10768-10776.

Gagnon-Bartsh, J.A.et al.(2012) Using control genes to correct for unwanted
variation in microarray data. Biostatistics, 13, 539-552.

Johnson, W.E.et al.(2007) Adjusting batch effects in microarray expression
data using Empirical Bayes methods. Biostatistics, 8, 118-127.

Leek, J.T.et al.(2007) Capturing Heterogeneity in Gene Expression Studies by
Surrogate Variable Analysis. PLoS Genetics, 3(9), e161

Wang, W.et al.(2003) Quantification of Proteins and Metabolites by Mass
Spectrometry without Isotopic Labeling or Spiked Standards. Anal. Chem., 75,
4818-4826.
}
