% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{ms_prepare}
\alias{ms_prepare}
\title{Prepare a mass spec quantification data frame for filtering, imputation,
and normalization. Also provides summaries of data structure (replicates,
subjects, group_by, etc.)}
\usage{
ms_prepare(data, subject_id = "subject_id", replicate = "replicate",
  abundance = "abundance", group_by = "spike", mz = "mz",
  rt = "rt", cvmax = 0.5, missing_val = 1,
  min_proportion_present = 1/3)
}
\arguments{
\item{data}{A tidy dataframe of quantification data.}

\item{subject_id}{Name (string) of the subject ID column.}

\item{replicate}{Name (string) of the replicate column. Set to NULL if no
replicates. TODO: test NULL.}

\item{abundance}{Name (string) of the abundance column.}

\item{group_by}{Name (string) of the group_by column.}

\item{mz}{Mass-to-charge ratio variable name.}

\item{rt}{Retention time variable name.}

\item{cvmax}{Acceptable level of coefficient of variation between replicates.}

\item{missing_val}{Value of missing data in the quantification data file.}

\item{min_proportion_present}{Minimum proportion present to summarize with
median or mean. Below this the compound will be set to 0.}
}
\value{
An `msprep` object with `stage(rtn) == "prepared"` containing
summarised quantification data, a dataset of compounds summarised by medians,
and other related summaries.
}
\description{
Function reads in raw data files and summarizes technical replicates as the
mean of observations for compounds found in 2 or 3 replicates and with
coefficient of variation below specified level, or median for those found in
3 replicates but excess CV.
}
\examples{

# Read in data file
data(msquant)

# Convert dataset to tidy format
tidy_data    <- ms_tidy(msquant, mz = "mz", rt = "rt")
prepped_data <- ms_prepare(tidy_data, replicate = "replicate")

# Or, using tidyverse/magrittr pipes 
library(magrittr)
prepped_data <- msquant \%>\% ms_tidy \%>\% ms_prepare

str(prepped_data)
str(prepped_data$summary_data)

}
