% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{ms_tidy}
\alias{ms_tidy}
\title{Function for converting wide mass spec quantification data into a tidy data
frame}
\usage{
ms_tidy(quantification_data, mz = "mz", rt = "rt",
  id_extra_txt = "Neutral_Operator_Dif_Pos_", separator = "_",
  id_spike_pos = 1, id_subject_pos = 2, id_replicate_pos = 3)
}
\arguments{
\item{quantification_data}{Data frame containing the quantification data.}

\item{mz}{Name of the column containing mass-to-charge ratios.}

\item{rt}{Name of the column containing retention time.}

\item{id_extra_txt}{Text to remove when converting quant variable names to
variables.}

\item{separator}{Character/string separating spike, subject and replicate
ids.}

\item{id_spike_pos}{Order in which the spike number occurs in the quant
variable names (after \code{id_extra_txt} is removed).}

\item{id_subject_id_pos}{Order in which the subject id occurs in the quant
variable names.}

\item{id_replicate_id_pos}{Order in which the replicate id occurs in the quant
variable names.}
}
\value{
A tidy data frame of quant data, with columns mz, rt, spike,
subject_id, replicate, and abundance.
}
\description{
Function reads in wide dataset of mass spec quantification data and converts
it to a tidy dataset.  This function assumes that the dataset is in a wide
format, with a column representing the retention time (rt), another
representing the mass-to-charge ratio (mz), and the remaining columns
containing MS quantification data.
}
\details{
It also assumes that the column names of quantification data start with some
consistent, informational but unnecessary text (id_extra_txt), and contain
the spike, subject ID, and replicate ID in a consistent position, all
separated by a consistent separator.

See \code{data(quantification)} for an example.  If your data doesn't fit
this format, view the function code for hints on tidying your data.  The core
of this function consists of \code{tidyr::gather()}, \code{dplyr::mutate()},
and \code{tidyr::separate()}.
}
\examples{

  # load raw abundance/quantification dataset
  data(msquant)
  # print original dataset (using tibble's nice printing)
  tibble::as_data_frame(msquant) 
  # tidy the dataset and print it
  ms_tidy(msquant, mz = "mz", rt = "rt")

}
