% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msquant_subject1.R
\docType{data}
\name{msquant_subject1}
\alias{msquant_subject1}
\title{Example mass spectrometry dataset.}
\format{
Data frame containing 2644 observations of 56 variables
\describe{
  \item{mz}{Mass-to-charge ratio}
  \item{rt}{Retention-time}
  \item{Neutral_Operator_Dif_Pos_1x_O1_A_01}{The remaining columns specify
  metabolite abundances found in each subject/spike-in/batch/replicate 
  combinations. Each columns begins with
  'Neutral_Operator_Dif_Pos' followed by the spike-in (1x, 2x, or 4x), then
  the batch (01, 02, or 03), the replicate (A, B, or C), and finally the
  subject ID (01 or 02), each seperated by '_'.}
}
}
\usage{
msquant_subject1
}
\description{
Data contains LC-MS samples for 1 subject, each with three features: 
spike-in (1x, 2x, 4x), batch (01, 02, 03), and technical replicate (A, B, C). 
The first two columns indicate mass-to-charge ratio and retention-time for
the 2644 unique metabolites observed in the samples. The remaining 54 
columns indicate metabolite abundance for each subject/spike-in/
batch/replicate combination. Note that this dataset is a single subject
version of the 'msquant' dataset provided by this package, made available
for quicker testing.
}
\examples{
  data(msquant_subject1)
}
\keyword{datasets}
