% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R
\name{readdata}
\alias{readdata}
\title{Function for importing raw data and summarizing replicates}
\usage{
readdata(clinical_data, quantification_data, link_data, cvmax = 0.5,
  missing = 1, linktxt, sean_extra_test_col = FALSE,
  sean_extra_test_row = FALSE, sean_exclude = NULL, name_list = NULL,
  id_list = NULL, id_replace_string = NULL)
}
\arguments{
\item{cvmax}{Acceptable level of coefficient of variation between replicates.}

\item{missing}{Value of missing data in the quantification data file.}

\item{linktxt}{Column name for Run ID field in Subject Link dataset}

\item{clinicalfile}{Name of the clinical data file.}

\item{quantificationfile}{Name of the data file containing the
quantification data.}

\item{linkfile}{Name of Subject link file.}
}
\value{
sum_data1 Matrix of summarized replicates, one obs per subject per
compound

clinical Clinical dataset

medians List of compounds that had excess CV and utilized the median
}
\description{
Function reads in raw data files and summarizes technical replicates as the
mean of observations for compounds found in 2 or 3 replicates and with
coefficient of variation below specied level, or median for those found in 3
replicates but excess CV.
}
\note{
Must have unique compounds in raw data file.  Raw quantification data
should have two columns titled mz and rt that are combined to make the
column header.
}
\examples{

  ### Read in data files
  quant    <- read.csv("./data-raw/Clinical.csv")
  link     <- read.csv("./data-raw/Quantification.csv")
  clinical <- read.csv("./data-raw/SubjectLinks.csv")

  ### Set variables for program
  cvmax   <- 0.5
  missing <- 1
  linktxt <- "LCMS_Run_ID"

  test <- readdata(clinical_data       = clinical,
                   quantification_data = quant,
                   link_data           = link,
                   cvmax               = 0.50,
                   missing             = 1,
                   linktxt             = linktxt)

  save(test, file = paste0("./data/test.Rdata"))

}
